Score variation fonts by distance to supported range
Before this change, every font was assumed to have a single
width and weight, and we computed the distance to that value
when finding a good font to render a CSS style.  After this
change, fonts can have a range of supported widths and weights,
and we compute the distance to the range for scoring.

+++ src/fontdiff/style.h
@@ -32,7 +32,7 @@ class Style {
   const Language* GetLanguage() const { return language_; }
   double GetFontSize() const { return fontSize_; }
   double GetFontWeight() const { return fontWeight_; }
-  int32_t GetFontScore(const Font& font) const;
+  double GetFontScore(const Font& font) const;
 
  private:
   void SetProperty(const std::string& key, const std::string& value);
@@ -41,7 +41,7 @@ class Style {
 
   const Language* language_;
   double fontSize_;
-  double fontWeight_;  // 100..900
+  double fontWeight_;  // 100..1000
 };
 
 }  // namespace fontdiff
